agent/requirements.txt
Flask==2.3.2
Flask_Cors==4.0.0
Flask_SocketIO==5.3.6
langchain==0.1.7
langchain_community==0.0.20
langchain_openai==0.0.6
python-dotenv==1.0.1
Requests==2.31.0
gunicorn==19.7.1